added working update site form

+++ src/app/components/site-list.tsx
@@ -1,23 +1,29 @@
 'use client'
+
 import { Fragment, useState } from 'react'
 import { Menu, Transition } from '@headlessui/react'
-import { IconDotsVertical } from '@tabler/icons-react'
+import { IconCircleArrowUpRight, IconDotsVertical, IconPencil, IconX } from '@tabler/icons-react'
 import { toast } from 'react-hot-toast'
 import Link from 'next/link'
 import { Site } from '@/types/site'
-
-function classNames(...classes: any) {
-  return classes.filter(Boolean).join(' ')
-}
 
 export interface SitesProps {
   sites: Site[]
 }
 
 export function SiteList({ sites: siteList }: SitesProps) {
+  
   const [sites, setSites] = useState(siteList)
+  const [isDeleting, setIsDeleting] = useState(false);
 
-  async function deleteSite(id: string) {
+  async function deleteSite(id: string, name:string) {
+
+    setIsDeleting(true);
+
+    var ok = confirm(`Are you sure you want to delete ${name}?`);
+
+    if(!ok) return;
+
     const response = await fetch(`/api/sites/${id}`, {
       method: 'DELETE',
     })
@@ -27,10 +33,11 @@ export function SiteList({ sites: siteList }: SitesProps) {
     if (!response.ok) {
       toast.error('There was an error while deleting you site.')
       console.error(data)
+    } else {
+      toast.success('Site deleted.')
     }
-    toast.success('Site deleted!')
     await getSites()
-    return data
+    setIsDeleting(false);
   }
 
   async function getSites() {
@@ -60,25 +67,25 @@ export function SiteList({ sites: siteList }: SitesProps) {
         >
           <div className="min-w-0">
             <div className="flex items-start gap-x-3">
-              <p className="text-sm font-semibold leading-6 text-neutral-300">
+              <p className="text-lg font-semibold leading-6 text-neutral-400">
                 {site.content.title}
               </p>
-            </div>
-            <div className="mt-1 flex items-center gap-x-2 text-xs leading-5 text-gray-500">
-              <p className="whitespace-nowrap">Deployed on</p>
-              <svg viewBox="0 0 2 2" className="h-0.5 w-0.5 fill-current">
-                <circle cx={1} cy={1} r={1} />
-              </svg>
             </div>
           </div>
           <div className="flex flex-none items-center gap-x-4">
             <a
               href={site.href}
               target="_blank"
-              className="hidden rounded-md bg-purple-600 px-2.5 py-1.5 text-sm font-semibold text-white shadow-sm ring-1 ring-inset ring-purple-800 hover:bg-purple-500 sm:block"
+              className="flex items-center rounded-md bg-purple-600 px-4 py-2 text-sm font-semibold text-white shadow-sm ring-1 ring-inset ring-purple-800 hover:bg-purple-500 gap-1"
             >
-              View site<span className="sr-only">, {site.bucketName}</span>
+              <span>View</span> <IconCircleArrowUpRight className='h-5 w-5'/> <span className="sr-only">, {site.bucketName}</span>
             </a>
+              <Link
+                href={`/sites/${site._id}`}
+                className='flex items-center rounded-md bg-purple-600 px-4 py-2 text-sm font-semibold text-white shadow-sm ring-1 ring-inset ring-purple-800 hover:bg-purple-500 gap-1'
+              >
+                <span>Edit</span> <IconPencil className='h-5 w-5'/> <span className="sr-only">, {site.bucketName}</span>
+              </Link>
             <Menu as="div" className="relative flex-none">
               <Menu.Button className="-m-2.5 block p-2.5 text-purple-400 hover:text-purple-700">
                 <span className="sr-only">Open options</span>
@@ -93,31 +100,17 @@ export function SiteList({ sites: siteList }: SitesProps) {
                 leaveFrom="transform opacity-100 scale-100"
                 leaveTo="transform opacity-0 scale-95"
               >
-                <Menu.Items className="absolute right-0 z-10 mt-2 w-32 origin-top-right rounded-md bg-white py-2 shadow-lg ring-1 ring-gray-900/5 focus:outline-none">
-                  <Menu.Item>
-                    {({ active }) => (
-                      <Link
-                        href={`/sites/${site._id}`}
-                        className={classNames(
-                          active ? 'bg-gray-50' : '',
-                          'block px-3 py-1 text-sm leading-6 text-gray-900',
-                        )}
-                      >
-                        Edit<span className="sr-only">, {site.bucketName}</span>
-                      </Link>
-                    )}
-                  </Menu.Item>
+                <Menu.Items className="absolute right-0 z-10 mt-2 w-28 origin-top-right rounded-md bg-white p-2 shadow-lg ring-1 ring-gray-900/5 focus:outline-none">
+                  
                   <Menu.Item>
                     {({ active }) => (
                       <button
                         type="button"
-                        onClick={() => deleteSite(site._id)}
-                        className={classNames(
-                          active ? 'bg-gray-50' : '',
-                          'block px-3 py-1 text-sm leading-6 text-gray-900',
-                        )}
+                        disabled={isDeleting}
+                        onClick={() => deleteSite(site._id, site.content.title)}
+                        className='disabled:opacity-40 disabled:cursor-wait rounded-md flex items-center justify-center gap-1 px-2 py-1 text-sm leading-6 text-white bg-red-600 w-full hover:bg-red-500'
                       >
-                        Delete
+                        Delete <IconX className='w-5 h-5'/>
                         <span className="sr-only">, {site.bucketName}</span>
                       </button>
                     )}
